requirements.txt
python-dotenv
jupyter
ipykernel
markdown
ratelimit
yapf
tldextract
pyperclip

openai
openai-agents
tiktoken
pydantic

numpy
pandas
scipy
hdbscan
umap-learn
choix

playwright>=1.43.0
playwright-stealth
aiofiles

pillow
matplotlib

feedparser
aiohttp
newsapi-python
trafilatura

chromadb

